v3 compute prep: stem hold-skip, dtype-safe parcel add, grad-noise GPU-accumulate
Three pre-varlen bf16/throughput cleanups (all bit-identical output):
- stem: skip repeat_interleave when factor==1 (native 32Hz hop=64) — it is NOT
  free, repeat_interleave(1) copies every band each step; clock guard preserved
  via the separate t32 check.
- towers: parcel_embed add casts .to(x.dtype) — nn.Embedding output isn't
  autocast-downcast, so under bf16-mixed the bare add promoted the residual
  stream to fp32. No-op under fp32.
- ssl_health_v3 grad_noise_scale: accumulate on-GPU float64 scalar, one .item()
  at the end instead of 2 blocking .item() per param tensor (hundreds of
  device->host stalls/step). Bit-identical (sequential double sum, same order).

87 v3 model + ssl_health tests pass.

+++ src/speech_decoding/models/v14_converged_v3/towers.py
@@ -106,7 +106,10 @@ class V3Tower(nn.Module):
         # capture (e.g. (3, 12) for the rankme/feat_std depth comparison). Empty ⇒
         # the default single-return path (no tuple), so non-monitor callers and the
         # compiled hot path are unchanged.
-        x = x + self.parcel_embed(parcel_id)[None, :, None, :]  # (B,N,T,d) + (1,N,1,d)
+        # (B,N,T,d) + (1,N,1,d). ``.to(x.dtype)``: nn.Embedding output is not autocast-
+        # downcast (stays fp32), so under bf16-mixed the bare add would promote the whole
+        # residual stream back to fp32; cast keeps it bf16. No-op under fp32.
+        x = x + self.parcel_embed(parcel_id).to(x.dtype)[None, :, None, :]
         taps: dict[int, Tensor] = {}
         for i, b in enumerate(self.blocks):
             x = b(x, geom, visible) if isinstance(b, L1Block) else b(x, visible)
